# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AETHERIA E-Commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: List products with filters
      parameters:
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: minPrice
          in: query
          schema:
            type: ["number", "null"]
        - name: maxPrice
          in: query
          schema:
            type: ["number", "null"]
        - name: color
          in: query
          schema:
            type: ["string", "null"]
        - name: size
          in: query
          schema:
            type: ["string", "null"]
        - name: inStock
          in: query
          schema:
            type: ["boolean", "null"]
        - name: onSale
          in: query
          schema:
            type: ["boolean", "null"]
        - name: featured
          in: query
          schema:
            type: ["boolean", "null"]
        - name: isNew
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── FAVORITES ─────────────────────────────────────────────────────────────
...
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: List orders (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ─── COUPONS ────────────────────────────────────────────────────────────────
...
  /promotions:
    get:
      operationId: getPromotions
      tags: [promotions]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Promotion"
    post:
      operationId: createPromotion
      tags: [promotions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromotionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Promotion"
...
  /promotions/{id}:
    patch:
      operationId: updatePromotion
      tags: [promotions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromotionUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Promotion"
    delete:
      operationId: deletePromotion
      tags: [promotions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── SHIPPING METHODS ──────────────────────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, price, categoryId]
      properties:
        id:
          type: integer
        name:
          type: string
        sku:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        shortDescription:
          type: ["string", "null"]
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        price:
          type: number
        salePrice:
          type: ["number", "null"]
        images:
          type: array
          items:
            type: string
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        stock:
          type: integer
        featured:
          type: boolean
        isNew:
          type: boolean
        active:
          type: boolean
    ProductInput:
      type: object
      required: [name, price, categoryId]
      properties:
        name:
          type: string
        sku:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        shortDescription:
          type: ["string", "null"]
        categoryId:
          type: integer
        price:
          type: number
        salePrice:
          type: ["number", "null"]
        images:
          type: array
          items:
            type: string
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        stock:
          type: integer
        featured:
          type: boolean
        isNew:
          type: boolean
        active:
          type: boolean
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        sku:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        shortDescription:
          type: ["string", "null"]
        categoryId:
          type: integer
        price:
          type: number
        salePrice:
          type: ["number", "null"]
        images:
          type: array
          items:
            type: string
        sizes:
          type: array
          items:
            type: string
        colors:
          type: array
          items:
            type: string
        stock:
          type: integer
        featured:
          type: boolean
        isNew:
          type: boolean
        active:
          type: boolean

    # ── Favorite ─────────────────────────────────────────────────────────────
...
    Order:
      type: object
      required: [id, status, total, createdAt]
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [pendiente, confirmado, pagado, enviado, entregado, cancelado]
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        province:
          type: string
        postalCode:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        discount:
          type: number
        shippingCost:
          type: number
        total:
          type: number
        couponCode:
          type: ["string", "null"]
        shippingMethodId:
          type: ["integer", "null"]
        paymentMethodId:
          type: ["integer", "null"]
        paymentMethodName:
          type: ["string", "null"]
        shippingMethodName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    OrderItem:
      type: object
      required: [id, productId, quantity, price]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        selectedSize:
          type: ["string", "null"]
        selectedColor:
          type: ["string", "null"]
        price:
          type: number
    OrderInput:
      type: object
      required: [sessionId, firstName, lastName, email, phone, address, city, province, postalCode, paymentMethodId, shippingMethodId]
      properties:
        sessionId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        province:
          type: string
        postalCode:
          type: string
        paymentMethodId:
          type: integer
        shippingMethodId:
          type: integer
        couponCode:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pendiente, confirmado, pagado, enviado, entregado, cancelado]

    # ── Coupon ────────────────────────────────────────────────────────────────
...
    Promotion:
      type: object
      required: [id, title, discountType, discountValue]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        discountType:
          type: string
          enum: [percentage, fixed]
        discountValue:
          type: number
        categoryId:
          type: ["integer", "null"]
        productId:
          type: ["integer", "null"]
        active:
          type: boolean
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        badge:
          type: ["string", "null"]
    PromotionInput:
      type: object
      required: [title, discountType, discountValue]
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        discountType:
          type: string
          enum: [percentage, fixed]
        discountValue:
          type: number
        categoryId:
          type: ["integer", "null"]
        productId:
          type: ["integer", "null"]
        active:
          type: boolean
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        badge:
          type: ["string", "null"]
    PromotionUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        discountType:
          type: string
        discountValue:
          type: number
        categoryId:
          type: ["integer", "null"]
        productId:
          type: ["integer", "null"]
        active:
          type: boolean
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        badge:
          type: ["string", "null"]

    # ── ShippingMethod ────────────────────────────────────────────────────────
...