# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AETHERIA E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  status: OrderStatus;
  firstName?: string;
  lastName?: string;
  email?: string;
  phone?: string;
  address?: string;
  city?: string;
  province?: string;
  postalCode?: string;
  items?: OrderItem[];
  subtotal?: number;
  discount?: number;
  shippingCost?: number;
  total: number;
  /** @nullable */
  couponCode?: string | null;
  /** @nullable */
  shippingMethodId?: number | null;
  /** @nullable */
  paymentMethodId?: number | null;
  /** @nullable */
  paymentMethodName?: string | null;
  /** @nullable */
  shippingMethodName?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}